+++ domain/src/main/java/com/weit/domain/repository/user/UserRepository.kt
@@ -12,7 +12,7 @@ interface UserRepository {
 
     suspend fun updateInformation(informationUpdateUser: User): Result<Unit>
 
-    suspend fun setUserId(userId: Int)
+    suspend fun setUserId(userId: Long)
 
-    suspend fun getUserId(): Int
+    suspend fun getUserId(): Long
 }
