+++ domain/src/main/java/com/weit/domain/usecase/user/SetUserIdUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class SetUserIdUseCase @Inject constructor(
     private val repository: UserRepository,
 ) {
-    suspend operator fun invoke(userId: Int) {
+    suspend operator fun invoke(userId: Long) {
         repository.setUserId(userId)
     }
 }
